+++ onyx-ai/src/main/kotlin/com/onyxdevtools/ai/transformation/ColumnTransform.kt
@@ -23,7 +23,7 @@ interface ColumnTransform : Serializable {
      *
      * @param values The input column to fit on.
      */
-    fun fit(values: DoubleArray) = Unit
+    fun fit(values: FloatArray) = Unit
 
     /**
      * Applies the transformation to the provided column.
@@ -31,7 +31,7 @@ interface ColumnTransform : Serializable {
      * @param values The input column to transform.
      * @return A new column with the transformation applied.
      */
-    fun apply(values: DoubleArray): DoubleArray = values
+    fun apply(values: FloatArray): FloatArray = values
 
     /**
      * Reverses the transformation previously applied to a column.
@@ -42,7 +42,7 @@ interface ColumnTransform : Serializable {
      * @param values The transformed column to invert.
      * @return A column representing the inverse transformation, if supported.
      */
-    fun inverse(values: DoubleArray): DoubleArray = values
+    fun inverse(values: FloatArray): FloatArray = values
 
     /**
      * Identify whether the transform has been fit
@@ -69,7 +69,7 @@ fun ColumnTransforms.fitAndTransform(matrix: Matrix): Matrix {
     for (c in 0 until cols) {
         val t = this[c] ?: continue
 
-        val col = DoubleArray(rows) { r -> result[r][c] }
+        val col = FloatArray(rows) { r -> result[r][c] }
         t.fit(col)
         val transformed = t.apply(col)
 
@@ -85,7 +85,7 @@ fun ColumnTransforms.apply(matrix: Matrix): Matrix {
 
     forEachIndexed { c, t ->
         if (t == null) return@forEachIndexed
-        val col = DoubleArray(rows) { r -> result[r][c] }
+        val col = FloatArray(rows) { r -> result[r][c] }
         val transformed = t.apply(col)
         for (r in 0 until rows) result[r][c] = transformed[r]
     }
@@ -99,7 +99,7 @@ fun ColumnTransforms.inverse(matrix: Matrix): Matrix {
 
     forEachIndexed { c, t ->
         if (t == null) return@forEachIndexed
-        val col = DoubleArray(rows) { r -> result[r][c] }
+        val col = FloatArray(rows) { r -> result[r][c] }
         val restored = t.inverse(col)
         for (r in 0 until rows) result[r][c] = restored[r]
     }
